Change project file structure
