+++ config.py
@@ -1,0 +1,16 @@
+AKCIE_BASE = 'http://mkck.sk/akciezoz.php?rok='
+DOCS_BASE = '/Users/danielkraic/work/code/web/mkck-old/unger/web/documents/akcie'
+AKCIE_MIMO_BASE = '/Users/danielkraic/work/code/web/mkck-old/unger/web/documents/akcie/2per/'
+
+YEAR_FILE = 'akciear.htm.txt'
+CONTENT_FILE = 'zapis.htm.txt'
+PHOTOS_FILE = 'foto.htm.txt'
+PHOTOS_DIR = 'fot'
+
+
+DEBUG = True
+
+
+without_photos = {
+    2015: [19]
+}
