+++ debug.py
@@ -1,0 +1,6 @@
+from config import DEBUG
+
+
+def debug(message: str) -> None:
+    if DEBUG:
+        print(message)
